PROF-8872: Run profiler integration tests on macOS and Windows (#3945)
* Run profiler integration tests on both macOS and Windows

* Exclude from Windows testing those tests aren't supposed to work on it.

* Make createSandbox Windows-friendly

* Increase profiler tests timeout as `yarn add` in sandbox creation takes a _lot_ of time.

* Disable failing OOM integration tests on Windows

* Only sync once

* Move profiler integration test out of general integration test suite so it will not run twice on Linux
